requirements.txt
alembic==1.18.0
annotated-types==0.7.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorlog==6.10.1
ConfigSpace==1.2.2
contourpy==1.3.3
cycler==0.12.1
dacite==1.9.2
dask==2025.12.0
dask-expr==2.0.0
dask-jobqueue==0.9.0
dill==0.4.0
distributed==2025.12.0
faiss-cpu==1.13.2
filelock==3.20.3
filetype==1.2.0
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.1.0
func_timeout==4.3.5
greenlet==3.3.0
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.45.1
locket==1.0.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.0
minify_html==0.18.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multimethod==1.12
networkx==3.6.1
numba==0.62.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
optuna==4.6.0
packaging==25.0
pandas==2.3.3
partd==1.4.2
patsy==1.0.2
phik==0.12.5
pillow==12.1.0
psutil==7.2.1
puremagic==1.30
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.9.0
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
sentence-transformers==5.2.0
setuptools==80.9.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.8.1
SQLAlchemy==2.0.45
statsmodels==0.14.6
stopit==1.1.2
sympy==1.14.0
tblib==3.2.2
threadpoolctl==3.6.0
tokenizers==0.22.2
toolz==1.1.0
torch==2.9.1
tornado==6.5.4
TPOT==1.1.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.6
triton==3.5.1
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
update-checker==0.18.0
urllib3==2.6.3
visions==0.8.1
Werkzeug==3.1.5
wikipedia==1.4.0
wordcloud==1.9.5
xgboost==3.1.3
ydata-profiling==4.18.1
zict==3.0.0
